Completely rewrite gui and sim files into entry.py
