+++ ev3sim/__main__.py
@@ -1,0 +1,3 @@
+from ev3sim.entry import main
+
+main()
