+++ ev3sim/batched_run.py
@@ -32,7 +32,7 @@ def simulate(batch_file, preset_filename, bot_paths, seed, override_settings, *q
     initialiseFromConfig(config, send_queues, recv_queues)
 
 
-def batched_run(batch_file, bind_addr, seed):
+def batched_run(batch_file, seed):
 
     with open(batch_file, "r") as f:
         config = yaml.safe_load(f)
